requirements.txt
anthropic>=0.40.0
streamlit>=1.35.0
python-dotenv>=1.0.0
anyio>=4.0.0
rich>=13.0.0
pydantic>=2.0.0
google-cloud-compute>=1.14.0
google-cloud-storage>=2.10.0
google-cloud-monitoring>=2.16.0
google-cloud-billing>=1.11.0
google-cloud-run>=0.10.0
google-cloud-resource-manager>=1.10.0
google-cloud-iam>=2.12.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
pyyaml>=6.0.0
slack-bolt>=1.18.0   # Phase 4 ChatOps — optional, only needed for --slack
